Refactor: move memory and cpu converter to the common module
