+++ k8spin_common/k8spin_common/resources/quotas.py
@@ -49,3 +49,31 @@ def create_resource_quota(api, name: str, namespace: str, labels: dict, cpu: str
     metadata = _obj.get("metadata")
     metadata["labels"] = labels
     return pykube.ResourceQuota(api, _obj)
+
+def memory_convert_unit(value: str) -> float:
+    # https://k8smeetup.github.io/docs/tasks/configure-pod-container/assign-cpu-ram-container/#cpu-and-ram-units
+    #E, P, T, G, M, K, Ei, Pi, Ti, Gi, Mi, Ki
+    value = value.replace("i", "")
+    scale_dict = {
+        "E": pow(10, 18),
+        "P": pow(10, 15),
+        "T": pow(10, 12),
+        "G": pow(10, 9),
+        "M": pow(10, 6),
+        "K": pow(10, 3)
+    }
+    for unit in scale_dict:
+        if unit in value:
+            value = value.replace(unit, "")
+            return float(value)*scale_dict[unit]
+    return float(value)
+
+
+def cpu_convert_unit(value: str) -> float:
+    # https://k8smeetup.github.io/docs/tasks/configure-pod-container/assign-cpu-ram-container/#cpu-and-ram-units
+    # 0.1, 1, 100m
+    scale = 1000
+    if "m" in value:
+        scale = 1
+        value = value.replace("m", "")
+    return float(value)*scale
